+++ tests/tests.rs
@@ -95,6 +95,7 @@ fn option_03() {
     assert_eq!(value.unwrap(), "test");
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_01() {
     let mut args = Arguments::from_vec(to_vec(&["-w=10"]));
@@ -102,6 +103,7 @@ fn eq_option_01() {
     assert_eq!(value.unwrap(), 10);
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_02() {
     let mut args = Arguments::from_vec(to_vec(&["-w='10'"]));
@@ -109,6 +111,7 @@ fn eq_option_02() {
     assert_eq!(value.unwrap(), 10);
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_03() {
     let mut args = Arguments::from_vec(to_vec(&["-w=\"10\""]));
@@ -116,6 +119,7 @@ fn eq_option_03() {
     assert_eq!(value.unwrap(), 10);
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_04() {
     let mut args = Arguments::from_vec(to_vec(&["--width2=15", "--width=10"]));
@@ -123,6 +127,7 @@ fn eq_option_04() {
     assert_eq!(value.unwrap(), 10);
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_01() {
     let mut args = Arguments::from_vec(to_vec(&["-w="]));
@@ -131,6 +136,7 @@ fn eq_option_err_01() {
                "the '-w' option doesn't have an associated value");
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_02() {
     let mut args = Arguments::from_vec(to_vec(&["-w='"]));
@@ -139,6 +145,7 @@ fn eq_option_err_02() {
                "the '-w' option doesn't have an associated value");
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_03() {
     let mut args = Arguments::from_vec(to_vec(&["-w=''"]));
@@ -147,6 +154,7 @@ fn eq_option_err_03() {
                "the '-w' option doesn't have an associated value");
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_04() {
     let mut args = Arguments::from_vec(to_vec(&["-w='\""]));
@@ -155,6 +163,7 @@ fn eq_option_err_04() {
                "the '-w' option doesn't have an associated value");
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_05() {
     let mut args = Arguments::from_vec(to_vec(&["-w='10\""]));
@@ -163,6 +172,7 @@ fn eq_option_err_05() {
                "the '-w' option doesn't have an associated value");
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_06() {
     let mut args = Arguments::from_vec(to_vec(&["-w-10"]));
@@ -170,12 +180,21 @@ fn eq_option_err_06() {
     assert_eq!(value.unwrap(), None);
 }
 
+#[cfg(feature = "eq-separator")]
 #[test]
 fn eq_option_err_07() {
     let mut args = Arguments::from_vec(to_vec(&["-w=a"]));
     let value: Result<Option<u32>, Error> = args.opt_value_from_str("-w");
     assert_eq!(value.unwrap_err().to_string(),
                "failed to parse 'a' cause invalid digit found in string");
+}
+
+#[cfg(not(feature = "eq-separator"))]
+#[test]
+fn no_eq_separator_01() {
+    let mut args = Arguments::from_vec(to_vec(&["-w=a"]));
+    let value: Result<Option<u32>, Error> = args.opt_value_from_str("-w");
+    assert_eq!(value.unwrap(), None);
 }
 
 #[test]